% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dep_var_mean.R
\name{compute_dep_var_mean}
\alias{compute_dep_var_mean}
\title{Compute the Mean of the Dependent Variable Across Specifications}
\usage{
compute_dep_var_mean(formulas, type = c("raw", "transformed"), data)
}
\arguments{
\item{formulas}{A list of model formulas.}

\item{type}{A character string indicating whether to compute the mean of the raw or transformed dependent variable. Options are "raw" or "transformed".}

\item{data}{The data frame containing the variables.}
}
\value{
A list of lines to be added to a regression table, with one mean per formula.
}
\description{
This version loops over the list of formulas and computes the mean
(raw or transformed) for each outcome variable as specified in the formula.
It then returns a formatted list of lines for the regression table.
}
