% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_label.R
\name{get_var_label}
\alias{get_var_label}
\title{Retrieve Variable Label from Term}
\usage{
get_var_label(term, data)
}
\arguments{
\item{term}{A character string representing the term for which the label is to be retrieved. This can be a variable name,
an interaction term (e.g., 'var1:var2' or 'var1*var2'), or a logarithmic transformation (e.g., 'log(var)' or 'log(1 + var)').}

\item{data}{The data frame containing the variables. Each variable can have a 'label' attribute that provides a descriptive label.}
}
\value{
A character string representing the label for the given term. If the term is an interaction, the labels of the
interacting variables are combined with ' : '. If the term involves a logarithmic transformation, the label is formatted accordingly.
}
\description{
This function retrieves the label of a variable from a given term and data frame. It handles interaction terms,
logarithmic transformations, and standard variable names. The function returns the appropriate label for the term.
}
