% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R
\name{get_labels}
\alias{get_labels}
\title{Retrieve Labels for Outcome Variable, Covariates, Fixed Effects, and Clustering Variables from a Formula}
\usage{
get_labels(formula, data)
}
\arguments{
\item{formula}{The model formula.}

\item{data}{The data frame containing the variables.}
}
\value{
A list with four elements: \code{dep_var_label} for the outcome variable label,
\code{covariate_labels} for the covariate labels, \code{fe_labels} for the fixed effects labels,
and \code{clustering_labels} for the clustering variable labels.
}
\description{
This function retrieves the label of the outcome variable, the labels of the covariates,
the fixed effects, and the clustering variables from a given formula and data frame.
}
